pyproject.toml
[tool.poetry]
name = "vertex-dashboards"
version = "0.1.0"
description = ""
authors = ["Dylan Baker <dylan@spectacles.dev>"]
readme = "README.md"
packages = [{ include = "vertex_dashboards", from = "src" }]

[tool.poetry.scripts]
start = "vertex_dashboards.main:main"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
pydantic = "^2.7.4"
uvicorn = "^0.30.1"
pre-commit = "^3.7.1"
google-cloud-storage = "^2.17.0"
google-cloud-logging = "^3.10.0"
google-cloud-aiplatform = "^1.56.0"
resend = "^2.1.0"
google-cloud-firestore = "^2.16.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.10"
mypy = "^1.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
mypy_path = "./src"
plugins = 'pydantic.mypy'
strict = true
namespace_packages = true
ignore_missing_imports = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = false
check_untyped_defs = true
no_implicit_optional = true
warn_unused_ignores = true

[tool.ruff]
line-length = 88  # Set maximum line length to 88 characters

[tool.ruff.lint]
extend-select = ["I"]  # Extend the selection to include import conventions

